backend/pyproject.toml
[tool.poetry]
name = "arabic-ocr-translation"
version = "0.1.0"
description = "Arabic text extraction and translation system"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"
python-multipart = "^0.0.6"
paddleocr = "^2.7.0"
pytesseract = "^0.3.10"
pymupdf = "^1.23.8"
pdfplumber = "^0.10.3"
python-docx = "^1.1.0"
pillow = "^10.1.0"
transformers = "^4.35.2"
torch = "^2.1.1"
sentencepiece = "^0.1.99"
langdetect = "^1.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true




